Added/Modified code for the link back to exhibit page
Moved a lot of the code that handles the link back to the
exhibit page(s) containing an item from item.php into
functions in CulCustomize. Also moved these functions into
the class definition, and made them static.

As far as the modifications are concerned, the link text now
contains the title of the exhibit page, as well as the parent
page. Also, added code to prevent duplicate printing of a link,
as well as short-circuiting code in case we arrived at the item
page without going through an exhibit page.