Test if need "also", and return if no exhibit pages
The code now tests if the word "also" is needed in the heading
to the links to exhibit page(s) - word is not needed if there is not a
link back to the exhibit page that lead the user to the item page.
Also, added a test that just returns if there are no exhibit pages
to link to.