Added code from admin/themes/default/css/style.css
In October 2013, code was added to
admin/themes/default/css/style.css to achieve the following:
-) Make headings smaller, decrease line spacing, increase image
size in the admin item show page.
-) Hide the delete button in the upper right corner of the
Edit Exhibit Screen because one may think it is used to delete
selected page, but it actually deletes the whole exhibition.
-) Color red the delete option under each exhibition in the
Browse Exhibit screen, to hightlight what it will do.