Define filter which will alphabetize list of collection in search form
The define filter is called by get_table_options in libraries/globals.php
The implementation below will alphabetize the list of collections.
We want this so that the options under "Search by Collection"
in the Search Items page (which we get when Advanced Search is
selected) will be in alphabetical order.
This filter also affects the display order of the collections in
the Edit Item page, and the Add an Item page.