Fix img title attribute for layouts with thumbnails
For the Exhibit Builder layouts with thumbnails, the value displayed
on mouseover of a thumbnail was the name of the file,
not the name of the item. For most browsers, mouseover will diplay
the value of the title attribute for the img element.
After the fix, both the title and alt attributes for the thumbnail
img contains the title of the associated item.