Add helper function to wrap call to files_for_item
The helper function basically just wraps a call to files_for_item.
Since all themes called files_for_items with the same arguments,
it is easier to wrap the function call in this helper function,
and have the themes call the helper function. In case we need to
add/change an argument to the files_for_item function call,
we just have to change it here, instead of having to do it in each theme.