Add a parameter to the cul_general_get_unit_* methods
This parameter, which defaults to NULL, allows us to specify the
library unit when the methods are called. If no argument is given,
it reverts to the old behavior, which looks at the theme option
which specifies the library unit.
However, since this function is now used in the CUL-version
of the Contribution plugin, which has no notion of a theme, the
library unit needs to be passed in.