Show owner of item in admin item page
The added code displays the username and full name for the owner
of an item. This information is only viewable in the admin item page,
and only if the logged-in user has an admin or superuser account.
If the item does not have an owner -- this can happen if the original
owner of the item was later deleted/removed from the list of Omeka
users -- the item page will display this fact.