Move highslide info from default for all files to JPEG-specific callback
The code before this change would add the highslide info (class and onclick
attributes) for all files. This works if we assume that most files will be images,
and that the PDF files will be handled by the PDF Embed plugin, and video and sound
will be handled by the Html5 Media plugin. However, with files now being
contributed through the Contribution plugin, it is possible that other types
of files will end up in Omeka items. So the code was changed to revert back
to the original default behavior, which was to make the files downloadable.
NOTE: This is a first stab at this, need to test all pages that may
include files. Also, probably need to have the JPEG callback handle other
types of images (like png?)