Change link to point to fullsize instead of original (highslide)
Highslide was having problems with tiff images that were uploaded:
The zomming was not working -- just said "LOADING ..."
The anchor that was wrapping the image and that contained the
highslide info was href'ing to the original image (tiff), instead
of the fullsize image, which is a jpeg version. So the href was
changed to point to the fullsize, and the zomming now works.