Link to original for jpeg, and fullsize for other types of images
For TIFFS (and possibly PNG) files, the tiff in the original subdir,
and the derivative (created when image is uploaded into Omeka) jpeg
in fullsize subdir are the same sized image and will expand to the
same size. However, for jpeg that are uploaded, the derivative in
the fullsize subdir are actually smaller (the subdir is mis-named,
as one of the Omeka developers commented). This explains why the
expanding works (in current code) for uploaded tiffs, but not
uploaded jpegs.
So, to recap: for highslide to work, link to the original for
jpeg's, and link to the fullsize for other types of images, for
example tiff's and png's.