Fix error "Trying to get property of non-object in" CulCustomize.php line 318
This error was first noticed in the error log file. During investigation, it was also noticed in a
few item pages inside a few exhibitions -- NOTE: not all item pages in these exhibitions were
generating the error, just a few. It turns out it was the item pages for items that were also
included in non-public exhibitions.
It seems that ExhibitPage::getExhibit() will return null if the exhibit
containing the exhibit page is not public AND the access is not being made via a logged-in
user with access to non-public exhibits.
Looked in the application code to verify the above theory, did not see actual code that
checks the privs of the logged-in user, if logged-in user there is. However, with the inheritance
chain and the ZEND framework, it is hard to figure out if restrictions are enforced at a lower
level than the app code.
Regardless, the three lines of code below (skip current iteration of $exhibit is null)
fixes the problem.
NOTE: getExhibit() is defined in the following file:
plugins/ExhibitBuilder/models/ExhibitPage.php
URLs that were generating the errors:
https://exhibitions.cul.columbia.edu/exhibits/show/fugitives/item/8982
https://exhibitions.cul.columbia.edu/exhibits/show/fugitives/item/8983
https://exhibitions.cul.columbia.edu/exhibits/show/realia/item/1525