Add processing of the cul-general option to hide link to item page
The cul-general theme will offer the option to hide the
"Click here for item information" link. This commit contains the
code changes to process this new option.
Note that other themes will not contain this option. In this case
the get_theme_option() method will return null (via an indirect  call
to Theme::getOption which returns null if an option is not defined).
In this case, the code will implement the default behavior, which is
to display the link.