Remove cul_exhibit_builder_attachment_markup, cul_exhibit_builder_thumbnail_gallery
The above two methods are not used anymore. They were used by the previous
version of Exhibit Builder to customize the display of items in an
exhibit page. This includes setting up the highslide info, as well as
creating the link to the item page.
The functionality is now implemented by, among others, filter(s) and
callback(s) defined in the CulCustomize plugin.