Change url in citation on exhibit item page to point back said page
Associated jira ticket: OMEKA-45
The definition of getCitation in models/Item.php sets the url given
in the citation to the url of the item record. However, we want it
to point to the item page within the exhibition. Since the citation
is being built within the view of the item page within the exhibition,
current_url will return the needed url. In the case of the generic
item record page (not within the context of an exhibition),
this will also work since the view will be the item record view, so
current_url will equal record_url