Add check for null return value from ExhibitPage::getExhibit(), take 2
getExhibit() method in the ExhibitPage class returns null if the exhibit
the page belongs to is not public AND the access is not being made via a
logged-in user with access to non-public exhibits, which is always
the case when the public visits our online exhibits.
The original commit, 93c2f0a5f2c591afd948253137343b5c6891ac82, did not
check for null in the code specific to legacy exhibitions.